requirements.txt
albumentations==1.3.0
decord==0.6.0
defusedxml==0.7.1
dlib==19.24.0
easydict==1.10
numpy ==1.23.3
omegaconf==2.2.3
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
PyYAML==6.0
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
timm==0.9.2
torch==1.13.0
torchscale==0.3.0
torchvision==0.14.0
tqdm==4.64.1
wandb==0.13.5
yacs==0.1.8
